// src/components/button/types.ts
export interface GlassButtonVariant {
  default: 'default';
  primary: 'primary';
  secondary: 'secondary';
  tertiary: 'tertiary';
  ghost: 'ghost';
  destructive: 'destructive';
  error: 'error';
  outline: 'outline';
  link: 'link';
  gradient: 'gradient';
  success: 'success';
  warning: 'warning';
}

export type GlassButtonVariantType = keyof GlassButtonVariant;